feat: remove over complicated json methods
As a result, the output timestamp will in nanoseconds instead of
seconds. This will break any integration with external tools like
playerctl. However, with the latest changes to the cli there is not any
reason to use playerctl. In case, user do need use playerctl than the
can convert player name and timestamps.